Fixes flake8 issues and test failures
