+++ decotools/tests/test_blob_extraction.py
@@ -1,10 +1,8 @@
 
 import pytest
-from collections import Counter
 import numpy as np
-from skimage.io import imsave
 
-from ..blob_extraction import is_hotspot, get_intensity_metrics
+from decotools.blob_extraction import is_hotspot
 
 
 def test_is_hotspot_threshold_1():
@@ -25,29 +23,3 @@ def test_is_hotspot_inconsistent_coords():
         is_hotspot(x, y)
     error = 'x_coords and y_coords must have the same shape.'
     assert error == str(excinfo.value)
-
-
-def test_get_intensity_metrics_no_files():
-    # Ensure that an empty input file list gives an empty DataFrame
-    df_metrics = get_intensity_metrics(files=[])
-    assert df_metrics.empty
-
-
-def test_get_intensity_metrics_bad_file():
-    # Ensure that non-existant input file raises IOError w/ reasonable message
-    with pytest.raises(IOError) as excinfo:
-        bad_file = '/this/file/is/not/here.png'
-        get_intensity_metrics(bad_file)
-    assert 'No such file' in str(excinfo.value)
-
-
-def test_get_intensity_metrics_columns(tmpdir):
-    # Create and save a test image to a temporary file
-    tmpfile = tmpdir.join('temp_image.png')
-    imsave(str(tmpfile), np.random.random((5, 5, 4)))
-
-    columns = ['max', 'mean', 'percentile_16', 'percentile_50',
-               'percentile_84']
-    df_metrics = get_intensity_metrics(str(tmpfile))
-
-    assert Counter(df_metrics.columns) == Counter(columns)
